Set up angular to use client-side routing.
I think we'd like to use routing to navigate from making a new
equation into the editor, and we will keep track of all the
equations we've made for you as well. For now we do this in a
local variable but in the future we will save it to the server.