Custom 404 page is working once deployed to Appspot, I need to finish the GUI for that page.
Saving a loading equations is fully functional.  There is a empty JSON Object
name equJSONObj created, which should be updated as the equation is edited.  This
is automatically saved to the datastore via AJAX when save is clicked.

When a user logs in, all equations in their history are loaded into the javascript
variable equationHistory, which is an array storing all of the JSON equation
objects that have been saved for that user.